Making jvmTest commonTest
