+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/DeleteTest.kt
@@ -14,32 +14,28 @@
 
 package io.zenoh
 
-import io.zenoh.prelude.KnownEncoding
 import io.zenoh.keyexpr.intoKeyExpr
-import io.zenoh.prelude.Encoding
+import io.zenoh.prelude.SampleKind
 import io.zenoh.sample.Sample
-import io.zenoh.value.Value
-import org.junit.jupiter.api.Test
+import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
-class PutTest {
+class DeleteTest {
 
     companion object {
-        const val TEST_KEY_EXP = "example/testing/keyexpr"
-        const val TEST_PAYLOAD = "Hello"
+        val TEST_KEY_EXP = "example/testing/keyexpr".intoKeyExpr().getOrThrow()
     }
 
     @Test
-    fun subscriber_receivesPutValue() {
+    fun subscriber_receivesDelete() {
         val session = Session.open().getOrThrow()
         var receivedSample: Sample? = null
-        val keyExpr = TEST_KEY_EXP.intoKeyExpr().getOrThrow()
-        session.declareSubscriber(keyExpr).with { sample -> receivedSample = sample }.res()
-        val value = Value(TEST_PAYLOAD.toByteArray(), Encoding(KnownEncoding.TEXT_PLAIN))
-        session.put(keyExpr, value).res()
+        session.declareSubscriber(TEST_KEY_EXP).with { sample -> receivedSample = sample }.res()
+        session.delete(TEST_KEY_EXP).res()
         session.close()
+
         assertNotNull(receivedSample)
-        assertEquals(value, receivedSample!!.value)
+        assertEquals(receivedSample!!.kind, SampleKind.DELETE)
     }
 }
