+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/KeyExprTest.kt
@@ -17,8 +17,6 @@ package io.zenoh
 import io.zenoh.exceptions.SessionException
 import io.zenoh.keyexpr.KeyExpr
 import io.zenoh.keyexpr.intoKeyExpr
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
 import kotlin.test.*
 
 class KeyExprTest {
@@ -124,7 +122,7 @@ class KeyExprTest {
         val keyExpr2 = "x/y/z".intoKeyExpr().getOrThrow()
         val undeclare2 = session.undeclare(keyExpr2).res()
         assertTrue(undeclare2.isFailure)
-        assertThrows<SessionException> { undeclare2.getOrThrow() }
+        assertTrue(undeclare2.exceptionOrNull() is SessionException)
 
         session.close()
         keyExpr.close()
